+++ app/src/main/kotlin/com/metrolist/music/extensions/CoroutineExt.kt
@@ -5,7 +5,6 @@
 
 package com.metrolist.music.extensions
 
-import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.collectLatest
@@ -28,5 +27,3 @@ fun <T> Flow<T>.collectLatest(
         collectLatest(action)
     }
 }
-
-val SilentHandler = CoroutineExceptionHandler { _, _ -> }