+++ app/src/main/kotlin/com/metrolist/music/lyrics/KuGouLyricsProvider.kt
@@ -6,15 +6,14 @@
 package com.metrolist.music.lyrics
 
 import android.content.Context
+import androidx.datastore.preferences.core.Preferences
 import com.metrolist.kugou.KuGou
 import com.metrolist.music.constants.EnableKugouKey
-import com.metrolist.music.utils.dataStore
-import com.metrolist.music.utils.get
 
 object KuGouLyricsProvider : LyricsProvider {
     override val name = "Kugou"
-    override fun isEnabled(context: Context): Boolean =
-        context.dataStore[EnableKugouKey] ?: true
+    override fun isEnabled(preferences: Preferences): Boolean =
+        preferences[EnableKugouKey] ?: true
 
     override suspend fun getLyrics(
         context: Context,